documentation: set up app explained
add all files needed to set up the app and keep it up and running

add files to backup the information
