+++ Set_up/Backup/pushGit.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+git add .
+git commit -m "$(date '+%Y-%m-%d %H:%M:%S')"
+git push originsSsh master
