+++ Set_up/Backup/transformData.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+# Remember to install node (apt install node -y)
+
+node /path/transformScript/transformProductos.js
+node /path/transformScript/transformCategorias.js
+node /path/transformScript/transformMarcas.js
+node /path/transformScript/transformOrdenes.js
+node /path/transformScript/transformUsers.js
