+++ Set_up/Docker/couchdb/etc/configuration.ini
@@ -1,0 +1,7 @@
+[chttpd]
+port = 5984
+bind_address = 0.0.0.0
+
+[httpd]
+enable_cors = true
+bind_address = 0.0.0.0
