+++ Set_up/Docker/couchdb/etc/docker.ini
@@ -1,0 +1,23 @@
+
+[admins]
+admin = -pbkdf2-26cb991546cdb73079a93c5a9adf2ee98084ac8f,b579d79d32da4d3e432ee1c1be34c48f,10
+
+[couchdb]
+uuid = a7ebad05ec58c6cbd6aefa79675f01d3
+
+[chttpd_auth]
+secret = babd0c48f11b0eea9796cec09fb4d9c8
+
+[chttpd]
+bind_address = 0.0.0.0
+port = 5984
+enable_cors = true
+
+[cluster]
+n = 1
+
+[cors]
+credentials = true
+origins = https://acosta-repuesto-cristian-ayala.vercel.app,https://acosta-repuesto.vercel.app
+headers = accept, authorization, content-type, origin, referer
+methods = GET, PUT, POST, HEAD, DELETE
