+++ Set_up/Docker/nginx/baremetalNginx.conf
@@ -1,0 +1,88 @@
+map $http_origin $allow_origin {
+	~^https:\/\/acosta-repuesto-[\w-]*[\w-]*(cristian-ayala)*\.vercel\.app$ $http_origin;
+	~^http:\/\/localhost:[0-9]+$ $http_origin;
+	default "";
+}
+
+upstream couchdb {
+  server couchdb:5984 fail_timeout=5s max_fails=0;
+}
+
+server {
+	listen 80;
+	listen [::]:80;
+	server_name www.couchdb.tk;
+	server_tokens off;
+
+	location /.well-known/acme-challenge/ {
+		root /var/www/certbot;
+	}
+
+	location / {
+			return 301 https://$host$request_uri;
+  }
+}
+
+server {
+	listen 443 ssl;
+	listen [::]:443;
+	server_name www.couchdb.tk; # managed by Certbot
+	ssl_certificate /usr/share/nginx/certificates/fullchain.pem;
+	ssl_certificate_key /usr/share/nginx/certificates/privkey.pem;
+	# include /etc/ssl-options/options-nginx-ssl.conf;
+	# ssl_dhparam /etc/ssl-options/ssl-dhparams.pem;
+	proxy_cookie_path / "/; Secure;HttpOnly; SameSite=none";
+
+	# Check if third party cookies are allowed
+	# The first third-party "JavaScript file" - served by nginx
+	location = /step1.js.php {
+		expires -1;
+		add_header Cache-Control "private";
+		etag off;
+		add_header Access-Control-Allow-Origin "*";
+		add_header Access-Control-Allow-Methods "*";
+		add_header Content-Type 'application/javascript; charset=UTF-8';
+		add_header Set-Cookie "third_party_c_t=hey there!;Max-Age=172000; Secure; SameSite=none";
+		return 200 'window._3rd_party_test_step1_loaded();';
+	}
+
+	# The second third-party "JavaScript file" - served by nginx
+	location = /step2.js.php {
+		add_header Access-Control-Allow-Origin "*";
+		add_header Access-Control-Allow-Methods "*";
+		add_header Content-Type 'application/javascript; charset=UTF-8';
+		set $test 'false';
+		if ($cookie_third_party_c_t = 'hey there!') {
+						set $test 'true';
+						# clear the cookie
+						add_header Set-Cookie "third_party_c_t=;expires=Thu, 01 Jan 1970 00:00:00 GMT; Secure; SameSite=none";
+		}
+		return 200 'window._3rd_party_test_step2_loaded($test);';
+	}
+
+	location / {
+		add_header 'Access-Control-Allow-Origin' $allow_origin;
+		add_header 'Access-Control-Allow-Headers' 'Authorization,Accept,Origin,DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range'; # always
+		add_header 'Access-Control-Allow-Methods' 'GET,POST,OPTIONS,PUT,DELETE,PATCH';
+		proxy_pass http://couchdb;
+		proxy_read_timeout 5m;
+		proxy_redirect off;
+		proxy_set_header Host $host;
+		proxy_buffering off;
+		proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+		proxy_set_header X-Forwarded-Ssl on;
+	}
+}
+
+
+
+
+# location / {
+#         proxy_pass http://localhost:3000;
+#         proxy_set_header X-Forwarded-Host $host;
+#         proxy_set_header X-Forwarded-Proto $scheme;
+#         proxy_http_version 1.1;
+#         proxy_set_header Upgrade $http_upgrade;
+#         proxy_set_header Connection "Upgrade";
+#         proxy_set_header Host $host;
+#     }
